refactoring: removed duplicate code
