+++ src/main/java/de/serosystems/lib1090/msgs/adsb/SurfacePositionV1Msg.java
@@ -1,5 +1,6 @@
 package de.serosystems.lib1090.msgs.adsb;
 
+import de.serosystems.lib1090.decoding.SurfacePosition;
 import de.serosystems.lib1090.exceptions.BadFormatException;
 import de.serosystems.lib1090.exceptions.UnspecifiedFormatError;
 import de.serosystems.lib1090.msgs.modes.ExtendedSquitter;
@@ -75,16 +76,7 @@ public class SurfacePositionV1Msg extends SurfacePositionV0Msg implements Serial
 	 */
 	@Override
 	public double getHorizontalContainmentRadiusLimit() {
-		switch (getFormatTypeCode()) {
-			case 0: return -1;
-			case 5: return 7.5;
-			case 6: return 25;
-			case 7:
-				return hasNICSupplementA() ? 75 : 185.2;
-			case 8:
-				return 185.2;
-			default: return -1;
-		}
+		return SurfacePosition.decodeHCR(getFormatTypeCode(), hasNICSupplementA());
 	}
 
 	/**
@@ -94,14 +86,7 @@ public class SurfacePositionV1Msg extends SurfacePositionV0Msg implements Serial
 	 */
 	@Override
 	public byte getNIC() {
-		switch (getFormatTypeCode()) {
-			case 0: case 8: return 0;
-			case 5: return 11;
-			case 6: return 10;
-			case 7:
-				return (byte) (hasNICSupplementA() ? 9 : 8);
-			default: return 0;
-		}
+		return SurfacePosition.decodeNIC(getFormatTypeCode(), hasNICSupplementA());
 	}
 
 	@Override
